+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 * Update to IDLeDOM 0.7.1.
 * Fix doctype creation and "significant whitespace" handling in DOMParser.
 * Ensure that Document::documentElement is always populated.
+* Ensure that Document::documentElement matches FilteredElementList where
+  appropriate.
 * EXPERIMENTAL: Added Node::getExtensionData() and Node::setExtensionData()
   methods to allow end-users to associate additional non-spec data off of
   Nodes.